requirements.txt
tensorflow==2.12.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.1
elasticsearch==8.7.0
tldextract==3.4.0
python-dotenv==1.0.0